+++ src/routes/AppRoutes.jsx
@@ -1,9 +1,10 @@
 import React from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { BrowserRouter, Routes, Route  } from "react-router-dom";
 import { Register, Info, Layout } from "../pages/client";
-import RegisterContract from "../pages/contract";
+import RegisterContract from "../pages/contract/Register";
 import EditClient from "../pages/client/Register/EditClient";
-
+import EditContract from "../pages/contract/Register/EditContract";
+import { LayoutContract } from "../pages/contract";
 // Aqui você pode adicionar as rotas do seu projeto
 // Você também pode adicionar rotas aninhadas e, se quiser, até dividir elas em arquivos diferentes
 // Exemplo: clientRoutes.js, adminRoutes.js, etc. Ai faz a importação e exporta tudo aqui
@@ -16,7 +17,10 @@ const AppRoutes = () => {
           <Route path=":id" element={<Info />} />
           <Route path=":id/editar" element={<EditClient />} />
         </Route>
-        <Route path="/contrato" element={<RegisterContract />} />
+        <Route path="/contrato" element={<LayoutContract   />}>
+          <Route index element={<RegisterContract />} />
+          <Route path=":id/editar" element={<EditContract />} />
+        </Route> 
       </Routes>
     </BrowserRouter>
   );
